Worked on project

+++ src/main/java/com/revature/service/MoonService.java
@@ -2,31 +2,32 @@ package com.revature.service;
 
 import java.util.List;
 
+import com.revature.MainDriver;
 import com.revature.models.Moon;
+import com.revature.models.Planet;
 import com.revature.repository.MoonDao;
+import com.revature.repository.PlanetDao;
 
 public class MoonService {
 
 	private MoonDao dao;
+	private PlanetDao planetDao = new PlanetDao();
+
 
 	public MoonService(MoonDao dao) {
 		this.dao = dao;
 	}
 
-	public List<Moon> getAllMoons() {
-		return dao.getAllMoons();
+	public List<Moon> getAllMoons(int currentUserId) {
+		return dao.getAllMoons(currentUserId);
 	}
 
-	public Moon getMoonByName(int myPlanetId, String moonName) {
-		Moon retrievedMoon = dao.getMoonByName(moonName);
-		if(retrievedMoon.getMyPlanetId() == myPlanetId)return retrievedMoon;
-		else return new Moon();
+	public Moon getMoonByName(String moonName, int currentUserId) {
+		return dao.getMoonByName(moonName,currentUserId);
 	}
 
-	public Moon getMoonById(int myPlanetId, int moonId) {
-		Moon retrievedMoon = dao.getMoonById(moonId);
-		if(retrievedMoon.getMyPlanetId() == myPlanetId)return retrievedMoon;
-		else return new Moon();
+	public Moon getMoonById(int moonId, int currentUSerId) {
+		return dao.getMoonById(moonId,currentUSerId);
 	}
 
 	public Moon createMoon(Moon m) {
@@ -35,17 +36,20 @@ public class MoonService {
 			if(retrievedMoon != null){
 				String retrievedMoonName = retrievedMoon.getName();
 				String mName = m.getName();
-				if(!mName.equals(retrievedMoonName))return dao.createMoon(m);
+				Planet retrievedPlanet = planetDao.getPlanetById(m.getMyPlanetId());
+				if(!mName.equals(retrievedMoonName) && retrievedPlanet.getOwnerId() == MainDriver.loggedInUserId)return dao.createMoon(m);
 			}
 		}
 		return new Moon();
 	}
 
-	public boolean deleteMoonById(int moonId) {
-		return dao.deleteMoonById(moonId);
+	public boolean deleteMoonById(int moonId, int currentUserId) {
+		Moon retrievedMoon = getMoonById(moonId,currentUserId);
+		if(retrievedMoon.getId() != 0)return dao.deleteMoonById(moonId);
+		else return false;
 	}
 
-	public List<Moon> getMoonsFromPlanet(int myPlanetId) {
-		return  dao.getMoonsFromPlanet(myPlanetId);
+	public List<Moon> getMoonsFromPlanet(int myPlanetId,int currentUserId) {
+		return  dao.getMoonsFromPlanet(myPlanetId,currentUserId);
 	}
 }
